linter fix, remove old filters
