--- src/components/SearchFilters/SearchFilterData-mock.js
@@ -1,116 +1,0 @@
-export default [
-	{
-		facet_section: "model",
-		facet_filters: [
-			{
-				facet_name: "Dataset available",
-				facet_column: "dataset_available",
-				facet_options: [
-					"dosing studies",
-					"mutation",
-					"cytogenetics",
-					"publication",
-					"patient treatment",
-					"copy number alteration",
-					"expression",
-				],
-			},
-			{
-				facet_name: "Datasource",
-				facet_column: "datasource",
-				facet_options: [
-					"Curie-OC",
-					"CMP",
-					"VHIO-BC",
-					"Wistar-MDAnderson-Penn",
-					"VHIO-CRC",
-					"HCMI",
-					"UOM-BC",
-					"DFCI-CPDM",
-					"TRACE",
-					"IRCC-CRC",
-					"IRCC-GC",
-					"MDAnderson",
-					"JAX",
-					"NKI",
-					"PDMR",
-					"VHIO-PC",
-					"HCI-BCM",
-					"Curie-BC",
-					"WUSTL",
-					"LIH",
-					"UMCG",
-					"UOC-BC",
-					"SJCRH",
-					"CRL",
-					"PIVOT",
-					"Curie-LC",
-					"PMLB",
-				],
-			},
-		],
-	},
-	{
-		facet_section: "molecular_data",
-		facet_filters: [
-			{
-				facet_name: "Cytogfenetics",
-				facet_column: "crytfogenetics",
-				facet_options: [
-					"ERBB2",
-					"PGR",
-					"KRT19",
-					"ROS1",
-					"MKI67",
-					"IDH1",
-					"KRT18",
-					"PTPRC",
-					"ESR1",
-					"ALK",
-				],
-				facet_example: "ESR1",
-			},
-			{
-				facet_name: "Breast cancer biaomarkers",
-				facet_column: "breast_cancer_biaomarkers",
-				facet_options: [
-					"PR/PGR positive",
-					"PR/PGR negative",
-					"HER2/ERBB2 positive",
-					"ER/ESR1 negative",
-					"HER2/ERBB2 negative",
-					"ER/ESR1 positive",
-				],
-			},
-			{
-				facet_name: "Cytogenetics",
-				facet_column: "crytogenetics",
-				facet_options: [
-					"ERBBa2",
-					"PGa",
-					"KRTa19",
-					"ROSa1",
-					"MKaI67",
-					"IDHa1",
-					"KRTa18",
-					"PTPaRC",
-					"ESRa1",
-					"ALaK",
-				],
-				facet_example: "ESR1",
-			},
-			{
-				facet_name: "Breast cancer biomarkers",
-				facet_column: "breast_cancer_biomarkers",
-				facet_options: [
-					"PR/PGR positaive",
-					"PR/PGR negataive",
-					"HER2/ERBB2 poasitive",
-					"ER/ESR1 negataive",
-					"HER2/ERBB2 neagative",
-					"ER/ESR1 positaive",
-				],
-			},
-		],
-	},
-];
